+++ sdk/experimental/tdf/reader.go
@@ -1,0 +1,5 @@
+// Experimental: This package is EXPERIMENTAL and may change or be removed at any time
+
+package tdf
+
+type Reader struct{}